Digital Answer Evaluation/requirements.txt
sqlalchemy
sentence-transformers
scikit-learn
nltk
python-dotenv
google-generativeai
torch